+++ app/build.gradle
@@ -72,4 +72,7 @@ dependencies {
     implementation 'com.google.android.gms:play-services-location:21.0.1'
     implementation 'com.google.android.libraries.places:places:2.7.0'
     implementation 'com.google.maps.android:android-maps-utils:2.3.0'
+    implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.5.1'
+    api "com.google.android.libraries.places:places:2.7.0"
+    implementation 'noman.placesapi:placesAPI:1.1.3'
 }
