+++ app/src/main/res/layout/activity_autocomplete_address.xml
@@ -16,116 +16,30 @@
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="Address Line 1"/>
+            android:text="검색" />
 
         <com.example.mukgoapplication.model.AutocompleteEditText
             android:id="@+id/autocomplete_address1"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:autofillHints=""
-            android:hint="Address Line 1"
+            android:hint="주소입력"
             android:imeOptions="actionNext"
-            android:inputType="text"/>
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Apartment, unit, suite, or floor #"/>
-
-        <EditText
-            android:id="@+id/autocomplete_address2"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:autofillHints=""
-            android:hint="Apartment, unit, suite, or floor #"
-            android:imeOptions="actionNext"
-            android:inputType="text"/>
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="City"/>
-
-        <EditText
-            android:id="@+id/autocomplete_city"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:autofillHints=""
-            android:hint="City"
-            android:imeOptions="actionNext"
-            android:inputType="text"/>
+            android:inputType="text" />
 
         <LinearLayout
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:orientation="horizontal"
-            android:baselineAligned="false">
-
-            <LinearLayout
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:orientation="vertical">
-
-                <TextView
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:text="State/Province"/>
-
-                <EditText
-                    android:id="@+id/autocomplete_state"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:autofillHints=""
-                    android:hint="State/Province"
-                    android:imeOptions="actionNext"
-                    android:inputType="textCapCharacters"/>
-
-            </LinearLayout>
-
-            <LinearLayout
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:orientation="vertical">
-
-                <TextView
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:text="Postal code"/>
-
-                <EditText
-                    android:id="@+id/autocomplete_postal"
-                    android:layout_width="match_parent"
-                    android:layout_height="0dp"
-                    android:layout_weight="1"
-                    android:autofillHints=""
-                    android:hint="Postal code"
-                    android:imeOptions="actionNext"
-                    android:inputType="number"/>
-            </LinearLayout>
-        </LinearLayout>
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Country/Region"/>
-
-        <EditText
-            android:id="@+id/autocomplete_country"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:autofillHints=""
-            android:hint="Country/Region"
-            android:imeOptions="actionNext"
-            android:inputType="text"/>
+            android:baselineAligned="false"/>
 
         <!-- Proximity check option provided for development testing convenience.
          User would typically not control this option. -->
-        <CheckBox android:id="@+id/checkbox_proximity"
+        <CheckBox
+            android:id="@+id/cbAddress"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="Check device proximity"/>
+            android:text="GPS사용" />
 
         <!-- The map for visual confirmation of the selected address -->
         <!-- Stub to only load the map after Autocomplete prediction selection -->
@@ -141,7 +55,7 @@
             android:id="@+id/autocomplete_save_button"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:text="Submit address"/>
+            android:text="save form" />
 
         <!-- Reset button provided for development testing convenience. Not recommended for user-facing
         forms due to risk of mis-click when aiming for Submit button. -->
