--- app/src/main/res/layout/activity_map_main.xml
@@ -1,66 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/layout_map_main"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".map.MapMainActivity">
-
-    <LinearLayout
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:tools="http://schemas.android.com/tools"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_margin="4dp"
-        android:orientation="vertical"
-        tools:context=".MainActivity">
-
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="Choose a color theme…" />
-
-        <Spinner
-            android:id="@+id/theme_spinner"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content" />
-
-        <Button
-            android:id="@+id/autocomplete_button"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="2dp"
-            android:text="Autocomplete"/>
-
-
-        <Button
-            android:id="@+id/autocomplete_address_button"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="2dp"
-            android:text="Autocomplete Address Form"/>
-
-        <Button
-            android:id="@+id/programmatic_autocomplete_button"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="2dp"
-            android:text="Programmatic Autocomplete + Geocoding"/>
-
-        <Button
-            android:id="@+id/place_and_photo_button"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="2dp"
-            android:text="Fetch Place and Photo"/>
-
-        <Button
-            android:id="@+id/current_place_button"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="2dp"
-            android:text="Current Place"/>
-
-    </LinearLayout>
-</androidx.constraintlayout.widget.ConstraintLayout>
